core/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
openai-agents>=0.0.3
litellm
httpx>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0
bcrypt>=4.1.2
PyJWT>=2.8.0
